requirements.txt
binance-connector==3.6.0
python-telegram-bot==21.6
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
aiohttp==3.9.5
tenacity==8.5.0